pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "masphd"
version = "0.1.0"
description = "PhD research codebase for train delay prediction and ensemble modelling"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Mas Gol" }
]

dependencies = [
    "numpy",
    "pandas",
    "scikit-learn==1.6.1",
    # "xgboost==3.1.3",
    # "matplotlib",
    # "seaborn",
    "pyyaml",
    "python-dotenv",
    "jupyter",
    "stomp.py>=8.0.0",
    # "PyXB~=1.2.6", # old and cause bugs, try to handle by myself
    "holidays"
]

[project.optional-dependencies]
dev = [
    # "pytest",
    # "black",
    # "ruff",
    "ipykernel",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
